feat(frienship): handle friendships on profile page

+++ src/main/java/com/borathings/borapagar/friendRequest/impl/FriendRequestControllerImpl.java
@@ -8,6 +8,8 @@ import com.borathings.borapagar.friendRequest.dto.FriendRequestUpdateDto;
 import com.borathings.borapagar.friendRequest.dto.response.FriendRequestResponseDto;
 import java.util.List;
 import java.util.Optional;
+
+import com.borathings.borapagar.user.UserEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -25,7 +27,7 @@ public class FriendRequestControllerImpl implements FriendRequestController {
     public ResponseEntity<Void> createFriendRequest(
             Authentication currentUser, @RequestBody FriendRequestCreateDto friendRequestCreateDto) {
         String fromUserLogin = currentUser.getName();
-        Integer toUserId = friendRequestCreateDto.toUserId();
+        long toUserId = friendRequestCreateDto.toUserId();
         friendRequestService.createFriendRequest(fromUserLogin, toUserId);
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
@@ -37,6 +39,7 @@ public class FriendRequestControllerImpl implements FriendRequestController {
         friendRequestService.updateFriendRequest(requestId, friendRequestUpdateDto.status());
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
+
 
     @Override
     public ResponseEntity<Void> deleteFriendRequest(Authentication authentication, Long friendRequestId) {
